øvrige triggere dukker opp for eøs-begrunnelser også
