--- src/schemas/baks/begrunnelse/ba-sak/nasjonaleTriggere/øvrigeTriggere.ts
@@ -1,20 +1,0 @@
-import { BegrunnelseDokumentNavn, SanityTyper } from '../../../../../util/typer';
-import { erEndretUtbetalingBegrunnelse } from './endringsårsakTrigger';
-import { vilkårTriggerTilMenynavn, øvrigeTriggertyper } from '../typer';
-import {
-  erNasjonalEllerInstitusjonsBegrunnelse,
-  lagUtfyltNasjonaltFeltMenFeilRegelverkRegel,
-} from '../utils';
-
-export const øvrigeTriggere = {
-  title: 'Øvrige triggere',
-  type: SanityTyper.ARRAY,
-  name: BegrunnelseDokumentNavn.ØVRIGE_TRIGGERE,
-  of: [{ type: SanityTyper.STRING }],
-  options: {
-    list: øvrigeTriggertyper.map(trigger => vilkårTriggerTilMenynavn[trigger]),
-  },
-  hidden: ({ document }) =>
-    erEndretUtbetalingBegrunnelse(document) || !erNasjonalEllerInstitusjonsBegrunnelse(document),
-  validation: rule => lagUtfyltNasjonaltFeltMenFeilRegelverkRegel(rule),
-};
